2.0.0 Changed to record for cadence seconds per camera
