+++ app_control.py
@@ -8,7 +8,7 @@ import json
 from datetime import datetime
 
 
-VERSION = '1.2.4'
+VERSION = '2.0.0'
 
 
 def writesettings():
@@ -32,9 +32,8 @@ def initialise():
                  'camera_controller2': 'http://192.168.0.21/control',
                  'camera_controller_timeout': 0.5,
                  'camera_frame_rate': 500,
-                 'camera_degrees': 45,
                  'camera_qty': 1,
-                 'camera_recMode': 'segmented',
+                 'camera_recMode': 'normal',
                  'camera_storage': 'sda1',
                  'camera_format': 'x264',
                  'sensor_debounce_time': 0.1}
